refactor(*): rename + deletion
