+++ build-logic/config/src/main/kotlin/corp/tbm/cleanwizard/buildLogic/config/CleanWizardDataClassGenerationPattern.kt
