--- foundation/codegen/universal/src/main/kotlin/corp/tbm/cleanwizard/foundation/codegen/universal/processor/DependencyInjectionFramework.kt
@@ -1,8 +1,0 @@
-package corp.tbm.cleanwizard.foundation.codegen.universal.processor
-
-enum class DependencyInjectionFramework {
-    NONE,
-    KOIN,
-    KOIN_ANNOTATIONS,
-    DAGGER;
-}
