--- processors/use-case/src/main/kotlin/corp/tbm/cleanwizard/processors/useCase/UseCaseProcessorFunctionType.kt
@@ -1,20 +1,0 @@
-package corp.tbm.cleanwizard.processors.useCase
-
-import kotlinx.serialization.SerialName
-import kotlinx.serialization.Serializable
-
-@Serializable
-sealed class UseCaseProcessorFunctionType {
-
-    @Serializable
-    @SerialName("Operator")
-    data object Operator : UseCaseProcessorFunctionType()
-
-    @Serializable
-    @SerialName("InheritRepositoryFunctionName")
-    data object InheritRepositoryFunctionName : UseCaseProcessorFunctionType()
-
-    @Serializable
-    @SerialName("CustomFunctionName")
-    class CustomFunctionName(val functionName: String) : UseCaseProcessorFunctionType()
-}
